fastapi-clean-example/pyproject.toml
[build-system]
requires = ["hatchling>=1.13"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[project]
name = "fastapi-clean-example"
version = "0.1"
description = "Practical Clean Architecture backend example built with FastAPI"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Ivan Borovets", email = "ivan.r.borovets@gmail.com" },
]
requires-python = "==3.13.*"
dependencies = [
    "alembic==1.17.1",
    "alembic-postgresql-enum==1.8.0",
    "bcrypt==5.0.0",
    "dishka==1.7.2",
    "fastapi-error-map==0.9.8",
    "fastapi==0.121.0",
    "orjson==3.11.4",
    "psycopg[binary]==3.2.12",
    "python-multipart==0.0.9",
    "pyjwt[crypto]==2.10.1",
    "sqlalchemy[mypy]==2.0.44",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "uuid-utils==0.11.1",
    "gherkin-official==29.0.0",
]

[dependency-groups]
dev = [
    "deptry==0.24.0",
    "import-linter==2.9",
    "mypy==1.17.0",
    "pre-commit==4.2.0",
    "ruff==0.12.5",
    "slotscheck==0.19.1",
    { include-group = "test" },
    { include-group = "uat" },
]
test = [
    "coverage==7.10.0",
    "httpx==0.28.1",
    "line-profiler==5.0.0",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
]
uat = [
    "behave>=1.2.6",
    "playwright>=1.50.0",
    "requests>=2.32.0",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "if __name__ == .__main__.:",
    "pass",
    "\\.\\.\\.",
    "from .* import .*",
    "import .*",
    "logging\\..*",
    "log\\..*",
    "@(abc\\.)?abstractmethod",
    ".* = NewType\\(.*\\)",
    ".* = Literal\\[.*\\]",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "**/__init__.py",
    "**/alembic/**",
]
concurrency = [
    "multiprocessing",
    "thread",
]
parallel = true
branch = true

[tool.deptry]
root = ["src"]
ignore = [
    "DEP002", # Unused dependencies
    "DEP003", # Transitive dependencies
]

[tool.importlinter]
root_package = "app"
exclude_type_checking_imports = true

[[tool.importlinter.contracts]]
id = "clean-architecture"
name = "inner must not import outer"
type = "layers"
containers = ["app"]
layers = [
    "(run)",
    "(setup)",
    "presentation",
    "infrastructure",
    "application",
    "domain",
]
ignore_imports = [
    "app.infrastructure.persistence_sqla.alembic.env -> app.setup.config.settings",
]

[tool.mypy]
mypy_path = ["src"]
files = [
    "config",
    "scripts",
    "src",
    "tests",
]
exclude = "^.*alembic.*$"
python_version = "3.13"
strict = true
pretty = true
follow_imports = "silent"
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
warn_no_return = true
warn_unreachable = true
plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin",
]

[tool.pydantic-mypy]
warn_required_dynamic_aliases = true
warn_untyped_fields = true
init_typed = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow"]
addopts = "-m 'not slow'"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = []

[tool.ruff]
line-length = 88
target-version = "py313"
preview = true  # experimental

[tool.ruff.format]
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
    "A", # flake8-builtins            https://docs.astral.sh/ruff/rules/#flake8-builtins-a
    "ARG", # flake8-unused-arguments  https://docs.astral.sh/ruff/rules/#flake8-unused-arguments-arg
    "ASYNC", # flake8-async           https://docs.astral.sh/ruff/rules/#flake8-async-async
    "B", # flake8-bugbear             https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "C4", # flake8-comprehensions     https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    "C90", # mccabe                   https://docs.astral.sh/ruff/rules/#mccabe-c90
    #    "COM", # flake8-commas            https://docs.astral.sh/ruff/rules/#flake8-commas-com
    # Incompatible with ruff formatter, but can be useful (uncomment once, then review changes)
    # See: https://github.com/astral-sh/ruff/issues/9216
    "DTZ", # flake8-datetimez         https://docs.astral.sh/ruff/rules/#flake8-datetimez-dtz
    "E", # pycodestyle-error          https://docs.astral.sh/ruff/rules/#error-e
    "ERA001", # commented-out-code    https://docs.astral.sh/ruff/rules/#eradicate-era
    "F", # pyflakes                   https://docs.astral.sh/ruff/rules/#pyflakes-f
    "FLY", # flynt                    https://docs.astral.sh/ruff/rules/#flynt-fly
    "I", # isort                      https://docs.astral.sh/ruff/rules/#isort-i
    "LOG", # flake8-logging           https://docs.astral.sh/ruff/rules/#flake8-logging-log
    "N", # pep8-naming                https://docs.astral.sh/ruff/rules/#pep8-naming-n
    "PERF", # Perflint                https://docs.astral.sh/ruff/rules/#perflint-perf
    "PL", # pylint                    https://docs.astral.sh/ruff/rules/#pylint-pl
    "PT", # flake8-pytest-style       https://docs.astral.sh/ruff/rules/#flake8-pytest-style-pt
    "PTH", # flake8-use-pathlib       https://docs.astral.sh/ruff/rules/#flake8-use-pathlib-pth
    "Q", # flake8-quotes              https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    "RET", # flake8-return (RET)      https://docs.astral.sh/ruff/rules/#flake8-return-ret
    "RSE", # flake8-raise (RSE)       https://docs.astral.sh/ruff/rules/#flake8-raise-rse
    "RUF", # Ruff-specific rules      https://docs.astral.sh/ruff/rules/#ruff-specific-rules-ruf
    "S", # flake8-bandit              https://docs.astral.sh/ruff/rules/#flake8-bandit-s
    "SIM", # flake8-simplify          https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
    "SLF", # flake8-self (SLF)        https://docs.astral.sh/ruff/rules/#flake8-self-slf
    "SLOT", # flake8-slots            https://docs.astral.sh/ruff/rules/#flake8-slots-slot
    "T20", # flake8-print             https://docs.astral.sh/ruff/rules/#flake8-print-t20
    "TCH", # flake8-type-checking     https://docs.astral.sh/ruff/rules/#flake8-type-checking-tch
    "TID", # flake8-tidy-imports      https://docs.astral.sh/ruff/rules/#flake8-tidy-imports-tid
    "UP", # pyupgrade                 https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "W", # pycodestyle-warning        https://docs.astral.sh/ruff/rules/#warning-w
]
ignore = [
    "N818", # error-suffix-on-exception-name
    "PLR0913", # too-many-arguments
    "PLR0917", # too-many-positional-arguments
    "PLR6301", # no-self-use
    "PTH123", # builtin-open
    "TC001", # typing-only-first-party-import
    "TC002", # typing-only-third-party-import
    "TC003", # typing-only-standard-library-import
    "TC006", # runtime-cast-value
    "UP015", # redundant-open-modes
]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true
split-on-trailing-comma = true

[tool.ruff.lint.per-file-ignores]
"src/app/infrastructure/persistence_sqla/alembic/**" = ["ALL"]
"tests/**" = [
    "ARG002", # unused-method-argument
    "PLC2801", # unnecessary-dunder-call
    "PLR2004", # magic-value-comparison
    "PT011", # pytest-raises-too-broad
    "S101", # assert
    "S106", # hardcoded-password-func-arg
    "S107", # hardcoded-password-default
]
#
"src/app/infrastructure/adapters/password_hasher_bcrypt.py" = ["E501"]  # line-too-long
"src/app/infrastructure/auth/handlers/constants.py" = ["S105"]          # hardcoded-password-string
"src/app/presentation/http/auth/constants.py" = ["S105"]                # hardcoded-password-string
"src/app/presentation/http/errors/translators.py" = ["ARG002"]          # unused-method-argument
"scripts/dishka/plot_dependencies_data.py" = ["T201"]                   # print

[tool.slotscheck]
strict-imports = true
exclude-modules = '''
^app\.infrastructure\.persistence_sqla\.alembic
'''
